fix issue written during refactor
